Add files via upload
As the CS50 guy says, its time to take the training wheels off and slowly start playing with the big boys.

My first python coded game done in VS code. Decently proud of it as I had to take a break from coding for awhile due to work, and figured this project would refresh my understanding of object orientated programming.